% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg}
\alias{pkg}
\title{using::pkg}
\usage{
pkg(package, min_version = NULL, repo = NULL, ...)
}
\arguments{
\item{package}{\code{{character | name}} package name}

\item{min_version}{\code{{package_version | character}} min package version}

\item{repo}{\code{{character}} repo url}

\item{...}{additional arguments to pass to \code{\link[base:library]{base::library()}}}
}
\value{
\code{{character | logical}}
attached packages, or boolean if \code{logical.return is TRUE}
}
\description{
Optionally add version constraints to library calls
}
\details{
This function is intended to always be called with the \code{using::} prefix. A
call without this prefix will throw a warning. This simplifies detecting
dependencies declared by using::pkg for use with \code{{renv}} and \code{{capsule}}
}
\examples{
\dontrun{
using::pkg(capsule)
using::pkg(dplyr, min_version = "0.8.0")
using::pkg(h3jsr, min_version = "1.0.0", repo = "https://github.com/obrl-soil/h3jsr")
}

}
