% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/using.R
\name{using}
\alias{using}
\title{using}
\usage{
using(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link[using:pkg]{using::pkg()}}}
}
\value{
\code{{character | logical}}
attached packages, or boolean if \code{logical.return is TRUE}
}
\description{
Optionally add version constraints to library calls
}
\details{
This is function is deprecated. See \code{\link[using:pkg]{using::pkg()}}.
}
\examples{
\dontrun{
using::pkg(capsule)
using::pkg(dplyr, min_version = "0.8.0")
using::pkg(h3jsr, min_version = "1.0.0", repo = "https://github.com/obrl-soil/h3jsr")
}

}
