% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/using.R
\name{using}
\alias{using}
\title{using}
\usage{
using(package, min_version = NULL, max_version = NULL, repo = NULL, ...)
}
\arguments{
\item{package}{\code{{character | name}} package name}

\item{min_version}{\code{{package_version | character}} min package version}

\item{max_version}{\code{{package_version | character}} max package version}

\item{repo}{\code{{character}} repo url}

\item{...}{additional arguments to pass to \code{\link[base:library]{base::library()}}}
}
\value{
\code{{character | logical}}
attached packages, or boolean if \verb{logical.return is TRUE}
}
\description{
Optionally add version constraints to library calls
}
\examples{
\dontrun{
using(magrittr)
using(h3jsr, min_version = "1.0.0", repo = "https://github.com/obrl-soil/h3jsr")
using("rlang", max_version = "0.4.0")
}

}
