% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/using.R
\name{using}
\alias{list_using}
\alias{stop_using}
\alias{using}
\title{Load objects from packages}
\usage{
using(...)

stop_using(...)

list_using()
}
\arguments{
\item{...}{names of objects to be loaded}
}
\description{
Loads objects contained in packages or the entire package
}
\details{
Given an object name, this function will make that object available
  in the current \link{search} path. If the name given names a package, the
  entire package will be added to the search path. If the name specifies an
  object within a package, only that object will be available in the search
  path. The namespace of the object's package will be loaded, so other
  objects from that package will be available if given a namespace prefix.
  Note that you can expose objects from packages that are not exported. This
  is generally a bad idea, except for purposes of debugging.

  Package namespaces are loaded by calling \code{\link{loadNamespace}} with
  default arguments. If you need more control, manage package loading
  manually prior to calling \code{using}.

  The function \code{stop_using} attempts to remove the object from the
  search path or an entire package. This is intrinsically less reliable for
  packages, so use with caution.
}
\examples{
using(using::use)
list_using()
formals(use)
stop_using(use)
list_using()
using(my_use = using::use)
list_using()
formals(my_use)
stop_using(my_use)

}

